TST: test_liquidmotor.py reorganization.
